solved day 11

+++ days/day_11/aoc_day_11_rust/src/lib.rs
@@ -5,7 +5,7 @@ pub enum OperationSign {
 }
 
 impl OperationSign {
-    pub fn apply(&self, item: u32) -> u32 {
+    pub fn apply(&self, item: u64) -> u64 {
         match self {
             OperationSign::PLUS(sub1, sub2) => sub1.resolve(item) + sub2.resolve(item),
             OperationSign::MUL(sub1, sub2) => sub1.resolve(item) * sub2.resolve(item),
@@ -16,11 +16,11 @@ impl OperationSign {
 #[derive(Debug)]
 pub enum SubstituteValue {
     ITEM,
-    LITERAL(u32),
+    LITERAL(u64),
 }
 
 impl SubstituteValue {
-    fn resolve(&self, item: u32) -> u32 {
+    fn resolve(&self, item: u64) -> u64 {
         match &self {
             SubstituteValue::ITEM => item,
             SubstituteValue::LITERAL(v) => *v,
